修改

+++ src/main/java/com/hg/inventory/modules/base/customer/service/impl/CustomerServiceImpl.java
@@ -1,15 +1,20 @@
 package com.hg.inventory.modules.base.customer.service.impl;
 
+import cn.hutool.core.collection.CollUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.hg.inventory.common.domain.vo.PageInfo;
 import com.hg.inventory.common.enums.DelFlagEnum;
+import com.hg.inventory.common.exception.ServiceException;
 import com.hg.inventory.modules.base.customer.domain.entity.Customer;
 import com.hg.inventory.modules.base.customer.domain.form.CustomerForm;
 import com.hg.inventory.modules.base.customer.mapper.CustomerMapper;
 import com.hg.inventory.modules.base.customer.service.CustomerService;
 import com.hg.inventory.modules.base.product.domain.entity.Product;
+import com.hg.inventory.modules.sales.domain.entity.SalesOrder;
+import com.hg.inventory.modules.sales.domain.form.SalesOrderForm;
+import com.hg.inventory.modules.sales.service.SalesOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +24,8 @@ import java.util.List;
 public class CustomerServiceImpl implements CustomerService {
     @Autowired
     private CustomerMapper customerMapper;
+    @Autowired
+    private SalesOrderService salesOrderService;
     @Override
     public Customer save(Customer customer) {
         int flag = 0;
@@ -49,6 +56,12 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public Boolean deleteById(Long id) {
+        SalesOrderForm salesOrderForm = new SalesOrderForm();
+        salesOrderForm.setCustomerId(id);
+        List<SalesOrder> salesOrders = salesOrderService.list(salesOrderForm);
+        if(CollUtil.isNotEmpty(salesOrders)){
+            throw new ServiceException("存在关联销售订单，无法删除");
+        }
         customerMapper.deleteById(id);
         return true;
     }
